+++ CHANGELOG.md
@@ -6,3 +6,8 @@
  
 - Added documentation to classes
 - Fixed some things to get more pub points ;) 
+
+## 1.1.0
+
+- Changed from using requests package which was outdated to just http
+- no more timeout parameter on get and post